requirements.txt
# Enhanced BTIA-AD Net Requirements
# For RTX 5070 Ti with CUDA 12.x

# Core
torch>=2.1.0
torchvision>=0.16.0

# BioMedCLIP
open_clip_torch>=2.20.0
transformers>=4.35.0

# Data
datasets>=2.14.0
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0

# Training
tqdm>=4.66.0
wandb>=0.15.0  # Optional: for logging

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
scikit-learn>=1.3.0
bert-score>=0.3.13
